+++ CHANGELOG.md
@@ -59,3 +59,4 @@
 - Enh #55: Introduce `Lists` class for `list-level` HTML rendering and related tests (@terabytesoftw)
 - Enh #56: Introduce `Table` class for `table-level` HTML rendering and related tests (@terabytesoftw)
 - Enh #57: Introduce `Template` helper class with tests (@terabytesoftw)
+- Bug #58: Rename `Element` and `BaseElement` to `Html` and `BaseHtml` for standards-compliant HTML rendering with tests (@terabytesoftw)